esg/services/extraction/pyproject.toml
[project]
name = "extraction"
version = "0.1.0"
description = "ESG Intelligence Platform - Extraction Service for processing company reports"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.47.0",
    "langchain>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
